Refactor Anthropic
