+++ council/llm/providers/__init__.py
@@ -1,3 +1,4 @@
+from .anthropic import AnthropicLLM, AnthropicLLMConfiguration
 from .gemini import GeminiLLM, GeminiLLMConfiguration
 from .groq import GroqLLM, GroqLLMConfiguration
 from .ollama import OllamaLLM, OllamaLLMConfiguration
